+++ id3tag/src/formats/images/mod.rs
@@ -10,7 +10,7 @@
 //! 6. Check if the file identified needs to be resized, and if so, save it with "-resized" appended to the filename
 //! 7. Load the file as the cover
 
-use image::{self, imageops::FilterType, ImageOutputFormat::Jpeg};
+use image::{self, imageops::FilterType, io::Reader as ImageReader, ImageFormat::Jpeg};
 use std::error::Error;
 use std::io::Cursor;
 
@@ -104,21 +104,30 @@ fn find_cover(
 /// Returns an error if the image cannot be read or if the aspect ratio is not within the expected range.
 /// The expected aspect ratio is within 1.5:1 and 1:1.5 (eg. 300x200, 200x300, 300x300, 200x200)
 pub fn read_cover(cover_file: &str, max_size: u32) -> Result<Vec<u8>, Box<dyn Error>> {
-    let img = image::open(cover_file)?;
+    let img = ImageReader::open(cover_file)?.decode()?;
 
     if !aspect_ratio_ok(img.width(), img.height()) {
         return Err(format!("Image {cover_file} is outside the expected ratio.").into());
     }
 
-    let mut eib = Cursor::new(Vec::new());
+    // Create a buffer to hold the image
+    let mut img_buffer: Cursor<Vec<u8>> = Cursor::new(Vec::new());
 
-    // TODO: Add tests to see if image size < max_size/2. If so, return an error that the image is too small.
+    // Check if the image is too small
+    let min_size: u32 = max_size / 2;
+    if (img.width() < min_size && img.height() < min_size) && max_size > 0 {
+        return Err(format!("Image {cover_file} is too small.").into());
+    }
+
+    // Resize the image if needed
     if (img.width() > max_size || img.height() > max_size) && max_size > 0 {
         let img_resized = img.resize(max_size, max_size, FilterType::Lanczos3);
-        img_resized.write_to(&mut eib, Jpeg(90)).unwrap_or_default();
+        img_resized
+            .write_to(&mut img_buffer, Jpeg)
+            .unwrap_or_default();
     } else {
-        img.write_to(&mut eib, Jpeg(90)).unwrap_or_default();
+        img.write_to(&mut img_buffer, Jpeg).unwrap_or_default();
     };
 
-    Ok(eib.into_inner())
+    Ok(img_buffer.into_inner())
 }