+++ CHANGELOG.md
@@ -30,6 +30,7 @@ All notable changes to this project will be documented in this file.
 ### Documentation
 
 - Changelog Update
+- Changelog update
 
 ### Feat